requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==5.0.6
django-cloudinary-storage==0.3.0
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
whitenoise==6.7.0
